+++ Tests/SwiftLintFrameworkTests/TrailingClosureRuleTests.swift
@@ -9,7 +9,7 @@ class TrailingClosureRuleTests: SwiftLintTestCase {
                 Example("offsets.sorted(by: { $0.offset < $1.offset })"),
                 Example("foo.something(0, { $0 + 1 })")
             ])
-            .with(triggeringExamples: [Example("foo.map({ $0 + 1 })")])
+            .with(triggeringExamples: [Example("foo.map(↓{ $0 + 1 })")])
 
         verifyRule(description, ruleConfiguration: ["only_single_muted_parameter": true])
     }